[CoreFoundation] Fix constructor call for default allocator.
The default allocator is actully just a NULL pointer, so we have to
call the constructor as if we already have a reference otherwise,
the constructor will try to take a reference and cause a crash.

CFType will not try to release the reference when it is disposed
because the Handle will already be IntPtr.Zero.